+++ src/main/java/com/ma_recruit/repository/community/CommentRepository.java
@@ -3,7 +3,5 @@ package com.ma_recruit.repository.community;
 import com.ma_recruit.entity.community.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigInteger;
-
-public interface CommentRepository extends JpaRepository<Comment, BigInteger> {
+public interface CommentRepository extends JpaRepository<Comment, Integer> {
 }
